Move get raw data to DataLoader
